requirements.txt
openai
PyMuPDF
Pillow
